Switch to GSAP scrolltrigger for BG color change
from intersection observer, as it doesn't trigger for custom thresholds with elements larger than viewport. The logic is "how much of an element is in view" and not "how much of the viewport is covered by an element".

This also creates expected dark-light BG transitions for the intro section on the homepage